requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.15
gunicorn==25.1.0
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.2.0
sqlparse==0.5.5
uritemplate==4.2.0
whitenoise==6.11.0
